requirements.txt
absl-py==0.7.1
actionlib==1.12.1
alabaster==0.7.12
angles==1.9.12
argh==0.26.2
argon2-cffi==20.1.0
asn1crypto==0.24.0
astor==0.7.1
async-generator==1.10
atari-py==0.1.7
attrs==20.2.0
Babel==2.6.0
backcall==0.2.0
beautifulsoup4==4.6.0
bleach==3.1.5
bondpy==1.8.5
Brlapi==0.6.6
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.4.12
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
cloudpickle==1.1.1
colorama==0.3.7
command-not-found==0.3
controller-manager==0.18.3
controller-manager-msgs==0.18.3
cryptography==2.1.4
cupshelpers==1.0
cv-bridge==1.13.0
cycler==0.10.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
dill==0.2.9
distro-info===0.18ubuntu0.18.04.1
docutils==0.14
dynamic-reconfigure==1.6.3
entrypoints==0.3
future==0.17.1
gast==0.2.2
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.14
gnupg==2.3.1
grpcio==1.20.1
gym==0.12.1
h5py==2.9.0
hidpidaemon==18.4.6
html5lib==0.999999999
httplib2==0.9.2
idna==2.10
image-geometry==1.13.0
imagesize==1.1.0
importlib-metadata==1.7.0
interactive-markers==1.11.5
ipykernel==5.3.4
ipympl==0.5.8
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.13.2
joint-state-publisher==1.12.15
joint-state-publisher-gui==1.12.15
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.1
jupyterlab-server==1.2.0
kdl-parser-py==1.13.1
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kernelstub==3.1.4
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
language-selector==0.1
laser-geometry==1.6.7
livereload==2.6.1
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Markdown==3.1
Markups==2.0.1
MarkupSafe==1.1.1
matplotlib==3.3.2
message-filters==1.14.10
mistune==0.8.4
mock==3.0.5
moveit-commander==1.0.7
moveit-ros-planning-interface==1.0.7
moveit-ros-visualization==1.0.7
mpi4py==3.0.1
mpmath==1.1.0
nbclient==0.5.0
nbconvert==6.0.2
nbformat==5.0.7
nest-asyncio==1.4.0
netifaces==0.10.4
notebook==6.1.4
numpy==1.19.2
olefile==0.45.1
opencv-python==4.1.0.25
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
port-for==0.3.1
progressbar2==3.39.3
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.7.1
psutil==5.6.2
ptyprocess==0.6.0
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.9
pycups==1.9.73
pydbus==0.6.0
pyenchant==2.0.0
pyglet==1.3.2
Pygments==2.7.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
PyOpenGL==3.1.0
pyparsing==2.4.7
pyRFC3339==1.0
pyrsistent==0.17.3
pysmbc==1.0.15.6
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
python-qt-binding==0.4.3
python-utils==2.3.0
python-xlib==0.20
pytz==2019.1
pyxdg==0.25
PyYAML==5.1
pyzmq==19.0.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-app==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qtconsole==5.0.1
QtPy==1.9.0
repoman==1.2.2
reportlab==3.4.0
requests==2.24.0
requests-unixsocket==0.1.5
resource-retriever==1.12.6
roman==2.0.0
rosbag==1.14.10
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.10
roslaunch==1.14.10
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.10
rosmake==1.14.9
rosmaster==1.14.10
rosmsg==1.14.10
rosnode==1.14.10
rosparam==1.14.10
rospkg==1.1.9
rospy==1.14.10
rosservice==1.14.10
rostest==1.14.10
rostopic==1.14.10
rosunit==1.14.9
roswtf==1.14.10
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-image-view==0.4.16
rqt-joint-trajectory-controller==0.17.2
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.2
rqt-py-console==0.4.8
rqt-reconfigure==0.5.3
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.13
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.13.16
scipy==1.3.0
scour==0.36
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Send2Trash==1.5.0
sensor-msgs==1.12.8
sessioninstaller==0.0.0
six==1.15.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinx-autobuild==0.7.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
srdfdom==0.5.2
ssh-import-id==5.7
sympy==1.4
system-service==0.3
systemd-python==234
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textile==3.0.0
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.10
tornado==6.0.4
tqdm==4.32.1
traitlets==4.3.3
ubuntu-drivers-common==0.0.0
ufw==0.36
urdfdom-py==0.4.5
urllib3==1.25.10
virtualenv==16.4.1
watchdog==0.9.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.5.1
xacro==1.13.9
xkit==0.0.0
zipp==3.1.0
